+++ core/auto-readme/src/data.ts
@@ -8,7 +8,7 @@ import { zod2md } from "zod2md";
 
 import type { AstComments } from "./comment";
 
-import { fileExists } from "./utilities";
+import { fileExists, loadCatalogs } from "./utilities";
 
 export type ActionData = Awaited<ReturnType<typeof loadActionData>>;
 
@@ -50,14 +50,17 @@ export async function loadActionData(
 					};
 				}
 
+				case "BADGE":
 				case "PKG": {
 					const inputPath = find("path");
 					const filename = inputPath
 						? path.resolve(path.dirname(file), inputPath)
 						: path.dirname(file);
 					const pkgJson = await readPackageJSON(filename);
+					const catalogs = await loadCatalogs();
 					return {
 						action: action.action,
+						catalogs: catalogs,
 						parameters: action.parameters,
 						pkgJson,
 					};