+++ core/remark-asciinema/example/index.js
@@ -1,6 +1,6 @@
 import { remark } from "remark";
 
-import asciinema from "../dist/index.mjs";
+import asciinema from "../dist/index.cjs";
 
 export async function pipeline() {
 	const file = await remark()