+++ core/catppuccin-opml/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stephansama/catppuccin-opml
 
+## 1.0.5
+
+### Patch Changes
+
+- fdf1800: updated auto-readme badge implementation
+
 ## 1.0.4
 
 ### Patch Changes