+++ core/prettier-plugin-handlebars/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stephansama/prettier-plugin-handlebars
 
+## 0.1.7
+
+### Patch Changes
+
+- fdf1800: updated auto-readme badge implementation
+
 ## 0.1.6
 
 ### Patch Changes